backend/requirements.txt
ai21==2.7.0
ai21-tokenizer==0.11.2
aiofiles==23.2.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.2.1
filelock==3.15.4
Flask==3.0.3
frozenlist==1.4.1
fsspec==2024.6.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.7
importlib_metadata==7.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
marshmallow==3.21.3
msgpack==1.0.8
mypy-extensions==1.0.0
openai==1.35.3
packaging==24.1
priority==2.0.0
protobuf==5.27.1
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
PyYAML==6.0.1
Quart==0.19.6
ray==2.30.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.18.1
sentencepiece==0.2.0
sniffio==1.3.1
taskgroup==0.0.0a4
tenacity==8.4.1
tokenizers==0.19.1
tomli==2.0.1
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
wsproto==1.2.0
zipp==3.19.2
